Remove dual mode support
