+++ server/src/main/java/io/crate/protocols/ssl/SslSettings.java
@@ -63,8 +63,7 @@ public final class SslSettings {
 
     public enum SSLMode {
         ON,
-        OFF,
-        DUAL;
+        OFF;
 
         static SSLMode parse(String value) {
             return switch (value.toLowerCase(Locale.ENGLISH)) {
@@ -73,7 +72,6 @@ public final class SslSettings {
                 case "true" -> ON;
                 case "off" -> OFF;
                 case "false" -> OFF;
-                case "dual" -> DUAL;
                 default -> throw new IllegalArgumentException(value + " is not a valid SSL mode setting");
             };
         }
